requirements.txt
cs50
flask
python-dotenv
schedule